requirements.txt

Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
flask-marshmallow==0.13.0
Flask-SQLAlchemy==2.4.4
python-dotenv==0.14.0